ethclient: move tests to separate package
This is done to make the RevertErrorData example code look better.
